Reverted to Prime Twitter4j
I had forked twitter4j to add the `unshortened_url` field to URL entities, but this field is actually added by twarc when downloading tweets, not by Twitter directly, so it seems incorrect to use it in a forked version of twitter4j. As such, I took this reference out and now rely on `expanded_url` from Twitter.

I also renamed the FindUsersByLink to ExtractUsersByLink to bring it into consistent use with the other extract classes.