Changed KLDiv Class to Accept Background Distribution as a Parameter
In the first version of this class, I had specified KLD's background distribution through a class member under the assumption that I would have a parent class for handling query expansion. This proved cumbersome and unnecessary since co-occurrence is so easily implemented standalone. Hence, I added the background parameter as a field for the expansion method directly. Should clean up some of the code as well.